move UnpauseFunction to new file
