+++ pkg/execution/state/redis_state/queue.go
@@ -1492,41 +1492,6 @@ func (q *queue) dropPartitionPointerIfEmpty(ctx context.Context, shard QueueShar
 	}
 }
 
-func (q *queue) UnpauseFunction(ctx context.Context, shardName string, acctID, fnID uuid.UUID) error {
-	shard, ok := q.queueShardClients[shardName]
-	if !ok {
-		return fmt.Errorf("invalid shard %q", shardName)
-	}
-
-	part := &QueuePartition{
-		ID:         fnID.String(),
-		FunctionID: &fnID,
-		AccountID:  acctID,
-	}
-
-	err := q.PartitionRequeue(ctx, shard, part, q.clock.Now(), false)
-	if err != nil && !errors.Is(err, ErrPartitionNotFound) && !errors.Is(err, ErrPartitionGarbageCollected) {
-		q.log.Error("failed to requeue unpaused partition", "error", err, "partition", part)
-		return fmt.Errorf("could not unpause partition: %w", err)
-	}
-
-	// Also unpause shadow partition if key queues enabled
-	if q.allowKeyQueues(ctx, acctID) {
-		shadowPart := &QueueShadowPartition{
-			PartitionID: fnID.String(),
-			FunctionID:  &fnID,
-			AccountID:   &acctID,
-		}
-		// requeue to earliest item
-		err = q.ShadowPartitionRequeue(ctx, shadowPart, nil)
-		if err != nil && !errors.Is(err, ErrShadowPartitionNotFound) {
-			return fmt.Errorf("could not unpause shadow partition: %w", err)
-		}
-	}
-
-	q.log.Trace("requeued unpaused partition", "partition", part.Queue())
-	return nil
-}
 
 func (q *queue) SetFunctionMigrate(ctx context.Context, sourceShard string, fnID uuid.UUID, migrateLockUntil *time.Time) error {
 	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "SetFunctionMigrate"), redis_telemetry.ScopeQueue)
